requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.3
django-multiselectfield==0.1.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
numpy==1.21.3
openpyxl==3.1.2
pandas==1.5.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
